assistant/requirements.txt
flask==2.3.3
python-dotenv==1.0.0
groq==0.22.0
requests==2.31.0
flask-sqlalchemy==3.1.1
gunicorn==21.2.0
flask-cors==4.0.0 